third_party/WebKit/LayoutTests/http/tests/devtools/geolocation-emulation-tests-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that geolocation emulation with latitude and longitude works as expected.


Running: testGeolocationUnavailable
Position unavailable

Running: testOverridenGeolocation
Latitude: 50 Longitude: 100

Running: testInvalidParam
error: Protocol Error: Invalid type of argument 'latitude' for method 'Emulation.setGeolocationOverride' call. It must be 'number' but it is 'boolean'.

Running: testInvalidGeolocation
Latitude: 50 Longitude: 100

Running: testNoOverride
Override was cleared correctly.
